[ui] Split time-context, provider and use-hook into separate files
