+++ app/src/app/types.d.ts
@@ -1,0 +1,3 @@
+import { Tables } from "@/lib/database.types";
+
+type Period = Tables<"relative_period_with_time">;
